BASH: find_and_jump: Prefer exact match (when only one single match)
When we want to search a directory and there is another directory whose
name is the "superstring" of the desired directory, find_and_jump does not
jump to the desired directory as there are multiple search hits.  For
example, searching for "spark" will return "pyspark" and "spark".
It is better to jump to the directory if it is a single match.

The new behaviour is especially helpful with BASH completions.  BASH
completions of the search term makes it easier input a search term that
match a full directory name.